requirements.txt
Django~=3.2
mysqlclient~=2.0
djangorestframework~=3.2
langchain~=0.1
langchain-community~=0.0
langchain_openai~=0.0
faiss-cpu~=1.7
openai~=1.7
unstructured~=0.11
tiktoken~=0.5
langchain_core~=0.1